refactor: rehome the test-only schema parser and delete its dead siblings
parse_geometry_type_from_schema is genuine test infrastructure (four test
files and the fixture generator verify written native geo types with it)
but had no production caller left after #829 - it now lives in
tests/geo_schema_helpers.py where coverage and vulture account for it
honestly. detect_geo_logical_type, _check_parquet_schema_string and
_check_extension_type had only their own unit tests keeping them alive
and are deleted with them.
