+++ tests/test_metadata_utils.py
@@ -1,158 +1,13 @@
 """Tests for core/metadata_utils.py module."""
 
-import pyarrow as pa
-
 from geoparquet_io.core.metadata_utils import (
     _calculate_overall_bbox,
-    _check_extension_type,
-    _check_parquet_schema_string,
-    detect_geo_logical_type,
-    parse_geometry_type_from_schema,
+    extract_bbox_from_row_group_stats,
+    format_geoparquet_metadata,
+    format_parquet_geo_metadata,
+    format_parquet_metadata_enhanced,
+    has_parquet_geo_row_group_stats,
 )
-
-
-class TestCheckParquetSchemaString:
-    """Tests for _check_parquet_schema_string function."""
-
-    def test_detects_geography(self):
-        """Test detection of Geography type in schema string."""
-        # Real Parquet schema format uses 'required group' or similar
-        schema_str = (
-            'required group geometry (Geography(Point, XY, crs={"type":"OGC","code":"CRS84"}))'
-        )
-        result = _check_parquet_schema_string("geometry", schema_str)
-        assert result == "Geography"
-
-    def test_detects_geometry(self):
-        """Test detection of Geometry type in schema string."""
-        schema_str = "required group geometry (Geometry(Point, XY))"
-        result = _check_parquet_schema_string("geometry", schema_str)
-        assert result == "Geometry"
-
-    def test_returns_none_for_no_geo_type(self):
-        """Test returns None when no geo type is present."""
-        schema_str = "optional binary name (STRING)"
-        result = _check_parquet_schema_string("name", schema_str)
-        assert result is None
-
-    def test_handles_special_characters_in_field_name(self):
-        """Test handling of special characters in field name."""
-        schema_str = "required group my.geometry (Geometry(Point, XY))"
-        result = _check_parquet_schema_string("my.geometry", schema_str)
-        assert result == "Geometry"
-
-
-class TestCheckExtensionType:
-    """Tests for _check_extension_type function."""
-
-    def test_returns_none_for_simple_type(self):
-        """Test returns None for non-extension type."""
-        field = pa.field("name", pa.string())
-        result = _check_extension_type(field)
-        assert result is None
-
-    def test_returns_none_for_binary_type(self):
-        """Test returns None for plain binary type."""
-        field = pa.field("geometry", pa.binary())
-        result = _check_extension_type(field)
-        assert result is None
-
-
-class TestDetectGeoLogicalType:
-    """Tests for detect_geo_logical_type function."""
-
-    def test_detects_geography_from_schema_string(self):
-        """Test detection of Geography from schema string."""
-        field = pa.field("geometry", pa.binary())
-        schema_str = "required group geometry (Geography(Point, XY))"
-        result = detect_geo_logical_type(field, schema_str)
-        assert result == "Geography"
-
-    def test_detects_geometry_from_schema_string(self):
-        """Test detection of Geometry from schema string."""
-        field = pa.field("geometry", pa.binary())
-        schema_str = "required group geometry (Geometry(Point, XY))"
-        result = detect_geo_logical_type(field, schema_str)
-        assert result == "Geometry"
-
-    def test_returns_none_for_non_geo_field(self):
-        """Test returns None for non-geometry field."""
-        field = pa.field("name", pa.string())
-        result = detect_geo_logical_type(field, None)
-        assert result is None
-
-    def test_detects_from_type_string(self):
-        """Test detection when type string contains Geography/Geometry."""
-        # Create a mock field with Geography in type string
-        field = pa.field("geometry", pa.binary())
-        # Without schema_str, it checks the type string
-        result = detect_geo_logical_type(field, None)
-        # Binary type doesn't have Geography/Geometry in type string
-        assert result is None
-
-
-class TestParseGeometryTypeFromSchema:
-    """Tests for parse_geometry_type_from_schema function."""
-
-    def test_parses_simple_geometry(self):
-        """Test parsing simple Geometry type."""
-        schema_str = "required group geometry (Geometry(Point, XY))"
-        result = parse_geometry_type_from_schema("geometry", schema_str)
-        assert result is not None
-        assert result.get("geometry_type") == "Point"
-        assert result.get("coordinate_dimension") == "XY"
-
-    def test_parses_geography_with_crs(self):
-        """Test parsing Geography type with CRS."""
-        schema_str = 'required group geometry (Geography(Polygon, XY, crs="OGC:CRS84"))'
-        result = parse_geometry_type_from_schema("geometry", schema_str)
-        assert result is not None
-        assert result.get("geometry_type") == "Polygon"
-        assert result.get("crs") == "OGC:CRS84"
-
-    def test_parses_geography_with_algorithm(self):
-        """Test parsing Geography type with algorithm."""
-        schema_str = "required group geometry (Geography(Point, XY, algorithm=spherical))"
-        result = parse_geometry_type_from_schema("geometry", schema_str)
-        assert result is not None
-        assert result.get("algorithm") == "spherical"
-
-    def test_parses_xyz_coordinate_dimension(self):
-        """Test parsing XYZ coordinate dimension."""
-        schema_str = "required group geometry (Geometry(LineString, XYZ))"
-        result = parse_geometry_type_from_schema("geometry", schema_str)
-        assert result is not None
-        assert result.get("geometry_type") == "LineString"
-        assert result.get("coordinate_dimension") == "XYZ"
-
-    def test_parses_multipolygon(self):
-        """Test parsing MultiPolygon geometry type."""
-        schema_str = "required group geometry (Geometry(MultiPolygon, XY))"
-        result = parse_geometry_type_from_schema("geometry", schema_str)
-        assert result is not None
-        assert result.get("geometry_type") == "MultiPolygon"
-
-    def test_returns_none_for_non_geo_field(self):
-        """Test returns None for non-geometry field."""
-        schema_str = "optional binary name (STRING)"
-        result = parse_geometry_type_from_schema("name", schema_str)
-        assert result is None
-
-    def test_handles_json_crs(self):
-        """Test handling of JSON CRS object."""
-        crs_json = '{"type": "PROJCRS", "name": "NAD83"}'
-        schema_str = f"required group geometry (Geometry(Polygon, XY, crs={crs_json}))"
-        result = parse_geometry_type_from_schema("geometry", schema_str)
-        assert result is not None
-        # CRS parsing may return dict or string depending on format
-        assert result.get("crs") is not None
-
-    def test_parses_geometry_collection(self):
-        """Test parsing GeometryCollection type."""
-        schema_str = "required group geometry (Geometry(GeometryCollection, XY))"
-        result = parse_geometry_type_from_schema("geometry", schema_str)
-        assert result is not None
-        assert result.get("geometry_type") == "GeometryCollection"
 
 
 class TestCalculateOverallBbox:
@@ -203,7 +58,6 @@ class TestHasParquetGeoRowGroupStats:
 
     def test_with_file_with_bbox_column(self, places_test_file):
         """Test with file that has bbox column."""
-        from geoparquet_io.core.metadata_utils import has_parquet_geo_row_group_stats
 
         result = has_parquet_geo_row_group_stats(places_test_file)
         # Places file should have bbox column
@@ -213,7 +67,6 @@ class TestHasParquetGeoRowGroupStats:
 
     def test_with_file_without_bbox(self, buildings_test_file):
         """Test with file that lacks bbox column."""
-        from geoparquet_io.core.metadata_utils import has_parquet_geo_row_group_stats
 
         result = has_parquet_geo_row_group_stats(buildings_test_file)
         assert isinstance(result, dict)
@@ -225,7 +78,6 @@ class TestExtractBboxFromRowGroupStats:
 
     def test_with_file_with_bbox_column(self, places_test_file):
         """Test extraction from file with bbox column."""
-        from geoparquet_io.core.metadata_utils import extract_bbox_from_row_group_stats
 
         result = extract_bbox_from_row_group_stats(places_test_file, "geometry")
         # Result depends on whether places file has proper bbox stats
@@ -237,7 +89,6 @@ class TestExtractBboxFromRowGroupStats:
 
     def test_with_file_without_bbox(self, buildings_test_file):
         """Test extraction from file without bbox column."""
-        from geoparquet_io.core.metadata_utils import extract_bbox_from_row_group_stats
 
         result = extract_bbox_from_row_group_stats(buildings_test_file, "geometry")
         # Should return None when no bbox column exists
@@ -248,8 +99,6 @@ class TestFormatGeoparquetMetadata:
     """format_geoparquet_metadata: terminal and JSON rendering of the geo key."""
 
     def test_terminal_output_places(self, places_test_file, capsys):
-        from geoparquet_io.core.metadata_utils import format_geoparquet_metadata
-
         format_geoparquet_metadata(places_test_file, json_output=False)
         out = capsys.readouterr().out
         assert "GeoParquet Metadata" in out
@@ -263,8 +112,6 @@ class TestFormatGeoparquetMetadata:
     def test_json_output_places(self, places_test_file, capsys):
         import json
 
-        from geoparquet_io.core.metadata_utils import format_geoparquet_metadata
-
         format_geoparquet_metadata(places_test_file, json_output=True)
         data = json.loads(capsys.readouterr().out)
         assert data["version"] == "1.0.0"
@@ -272,8 +119,6 @@ class TestFormatGeoparquetMetadata:
         assert data["columns"]["geometry"]["encoding"] == "WKB"
 
     def test_terminal_no_geo_metadata(self, capsys):
-        from geoparquet_io.core.metadata_utils import format_geoparquet_metadata
-
         # crs-projjson.parquet is plain Parquet with no 'geo' key.
         format_geoparquet_metadata("tests/data/crs-projjson.parquet", json_output=False)
         out = capsys.readouterr().out
@@ -281,8 +126,6 @@ class TestFormatGeoparquetMetadata:
 
     def test_json_no_geo_metadata(self, capsys):
         import json
-
-        from geoparquet_io.core.metadata_utils import format_geoparquet_metadata
 
         format_geoparquet_metadata("tests/data/crs-projjson.parquet", json_output=True)
         assert json.loads(capsys.readouterr().out) is None
@@ -292,8 +135,6 @@ class TestFormatParquetGeoMetadata:
     """format_parquet_geo_metadata: Parquet-spec geospatial metadata section."""
 
     def test_terminal_no_native_geo_columns(self, places_test_file, capsys):
-        from geoparquet_io.core.metadata_utils import format_parquet_geo_metadata
-
         # places is GeoParquet 1.0 WKB: no native Parquet geo logical types.
         format_parquet_geo_metadata(places_test_file, json_output=False)
         out = capsys.readouterr().out
@@ -301,8 +142,6 @@ class TestFormatParquetGeoMetadata:
         assert "No geospatial columns detected" in out
 
     def test_terminal_native_geometry_column(self, capsys):
-        from geoparquet_io.core.metadata_utils import format_parquet_geo_metadata
-
         # GeoParquet 2.0 fixture with a native Geometry column and PROJJSON CRS.
         format_parquet_geo_metadata("tests/data/fields_gpq2_5070_brotli.parquet", json_output=False)
         out = capsys.readouterr().out
@@ -311,8 +150,6 @@ class TestFormatParquetGeoMetadata:
 
     def test_json_native_geometry_column(self, capsys):
         import json
-
-        from geoparquet_io.core.metadata_utils import format_parquet_geo_metadata
 
         format_parquet_geo_metadata("tests/data/fields_gpq2_5070_brotli.parquet", json_output=True)
         data = json.loads(capsys.readouterr().out)
@@ -328,8 +165,6 @@ class TestFormatParquetMetadataEnhanced:
     """format_parquet_metadata_enhanced: full Parquet file metadata section."""
 
     def test_terminal_output(self, places_test_file, capsys):
-        from geoparquet_io.core.metadata_utils import format_parquet_metadata_enhanced
-
         format_parquet_metadata_enhanced(
             places_test_file, json_output=False, primary_geom_col="geometry"
         )
@@ -342,8 +177,6 @@ class TestFormatParquetMetadataEnhanced:
     def test_json_output(self, places_test_file, capsys):
         import json
 
-        from geoparquet_io.core.metadata_utils import format_parquet_metadata_enhanced
-
         format_parquet_metadata_enhanced(places_test_file, json_output=True)
         data = json.loads(capsys.readouterr().out)
         assert data["num_rows"] == 766
@@ -353,8 +186,6 @@ class TestFormatParquetMetadataEnhanced:
 
     def test_json_output_all_row_groups(self, places_test_file, capsys):
         import json
-
-        from geoparquet_io.core.metadata_utils import format_parquet_metadata_enhanced
 
         format_parquet_metadata_enhanced(places_test_file, json_output=True, row_groups_limit=None)
         data = json.loads(capsys.readouterr().out)
